requirements.txt
streamlit
groq
PyPDF2
python-docx
gtts
openai-whisper
